Refactor challenge response functionality.
